crash.subsystem.storage: Cleanup singlequeue definitions
Rename function definitions in blocksq.py to start with sq prefix so
that we can have also corresponding mq alternatives without clashing
names. Also move request_age_ms to common part since it is not sq
specific.
